Fix last CR comments
